move biomes definition to a json file
