+++ res/generators/default.files/biomes.json
@@ -1,0 +1,11 @@
+{
+    "flat": {
+        "parameters": [],
+        "layers": [
+            {"height": -1, "block": "core:obstacle"}
+        ],
+        "sea_layers": [
+            {"height": -1, "block": "core:obstacle"}
+        ]
+    }
+}
