+++ src/content/ContentBuilder.cpp
@@ -89,7 +89,7 @@ std::unique_ptr<Content> ContentBuilder::build() {
     }
 
     for (auto& [name, def] : content->generators.getDefs()) {
-        def->script->prepare(*def, content.get());
+        def->prepare(content.get());
     }
 
     return content;
